Fix dfn.solve call high thread usage
Previously I didn't specify a method to solve the matrices in the `solve` call
so FEniCS would automatically use a huge number of threads (no idea
why). I specified using the 'petsc' solve in both the adjoint and
forward codes which leads to a nice performance boost and only one
core being used.