Fix adjoint to only PETSc vectors
Previously the gradient in adjoint.py was a numpy array. This meant that
everytime the adjoint calculation added a PETSc vector to the gradient,
a conversion from PETSc to numpy was done. This adds a lot of overhead.

I fixed this so that adjoint.py uses the gradient as a PETSc vector,
which has sped up the code by a decent amount.