Debug stuff
I think I've narrowed down the cause of problems in the adjoint solver.
When the separation point/minimum area changes tiny errors are induced
in the adjoint solution that ultimately affect the gradient. These
errors are sometimes more noticeable over different separation
instances so it will be tricky to figure out what aspect exactly is
causing the problem and how to fix it.