Continue debugging adjoint convergence issues
There are two interesting things going on with the adjoint thing now.
1. Increasing the collision midline offset can have large implications
for the convergence of the gradient. For example, changing from 0.002 to
0.005 or 0.01 leads to poor convergence. I don't have many ideas on why
this is.

2. Increasing the degree of the minimum smoothness parameter also has a
large impact on the convergence of the gradient. This could be due to
numerical overflow or underflow? I also remember in some cases that
increasing the 'sharpness' of the smooth maximum parameter could offset
the convergence issues that happened when the midline collision offset
was increased. This needs further investigation.