Remove TimingProperties and refactor StateFile
Removed TimingProperties as a class because it doesn't really fit the
way it's programmed now. Currently the Properties are only for
properties that are totally defined from a mesh while timing properties
is more free to be specific. Now it's just passed as a dictionary.

I also change StateFile to take a model as a parameter since
FluidProperties and SolidProperties now take model to initialize.