Refactor ForwardModel to Solid and Fluid models
Refactored a lot of the code mainly so that the forward model is now
composed of a solid and fluid model. This makes the forward model class
a lot less overloaded. I had to change a whole bunch of things because
of this so there are probably some things that are still broken.

I also made it so that properties are defined where the models are
defined.