Add 'safe' area to fluid model
Added a 'safe' area to the fluid model so that areas are restricted to a
lower bound. This prevents strange and undefined behaviour of the fluid
model as the area function for each surface vertex approaches zero.

Examples of this strange behaviour include:
- Negatives areas of fluid are undefined for the pressure/flow rate
- If the minimum area is close to zero, even small deviations in area
relative to the minimum can correspond to large drops in pressure. For
example, if a_sep=0.001 cm and a_i=0.0005 cm at a vertex downstream of
the separation point, the pressure at i will be a large negative value.
This is because the flow rate at a_i is 0.5 of the flow rate at a_sep
which induces a large drop in pressure from the Bernoulli effect. This
problem tends to happend near zero areas/collision because even small
relative displacements of the solid vertices correspond to large
ratios of areas.