Refactor dp -> dsolid, dfluid, dt0, ddt
Functionals now have derivatives w.r.t solid, fluid, t0, and dt. This is
consistent with the way the forward function is called and the set of
parameters needed to fully specify a step of the forward model.

`dsolid` was implemented in `adjoint`; now `ddt` and `dt0` need to be
fixed too.