Fix functional and adjoint to use dsolid, dfluid, etc
Functional was updated to use dsolid, dfluid, ddt, dt0 in place of the
single dp (for parameters) method. This better represents the standard
inputs for fully defining one iteration of the forward model. The
adjoint code was updated to use these new methods, except for dt0 and
dfluid mainly due to laziness and the fact that those derivatives
haven't been used yet.

Some other fixes were also applied to some of the functionals. For
example, I changed the eval() method for some, and fixed the eval_duva,
etc. methods for others that weren't updated before.