Fix df_dfp assembly in implicit newton solver
The implicit newton solver previously outlined had no df_dp sensitivity
matrix assembled. I added code to assemble this and also added a matrix
column permutation function. This function is needed to transform the
matrix df_dp from the 'p' ordering on the solid model, to the equivalent
p ordering on the fluid model.

There is still something wrong in the column permutation function. It is
throwing an error since it's trying to map all nonzero columns, even
if their not included in the permutation.