Fix newton method for implicit coupling
The newton method for solving the implicitly coupled FSI model is now
working. The method takes fewer iterations than the fixed point style
one, but is actually slower. It seems that the block matrix formation
commands end up taking a long time which makes the whole thing a bit
slower in the end. If you improve the block matrix assembly routines it
should be quite a bit faster.