Add better caching behaviour to StateFile
Reading state histories from the HDF5 file was one of the slowest
operations in computing the functional/adjoint.

This commit adds caching behaviour to chunks of h5 data so that the
number of reads required to process a file is decreased a lot. This
speeds up evaluation of functionals and the adjoint, sometimes by 2
orders of magnitude.