Implement functional operator overloading logic and bug fixes
Implemented some basic logic for operator overloading of functionals.

Also fixed a bunch of bugs in the StateFile reading/writing that were
leading to weird behaviour depending on the mode the file is opened
with. There are likely still more bugs with the StateFile as I didn't
think it through very well.