Refactor Solid set_state to single argument
Changed the Solid class set_state style methods to use a signle argument
uva rather than u, v, a. This is more consistent with the methods that
set the whole system behaviour like state, parameters, etc.

I have made a best attempt to refactor the rest of the code in
accordance with the change but might have missed some things.