Refactor fluid/solid properties to use BlockVec
This refactor has potential to make some of the code much simpler as you no
longer have to manually implement block addition, etc. everywhere.

There is still more refactoring to do as not all parts of the code are
using BlockVecs consistently so the refactored code does not play nice
with each other yet.