Update fluids to better handle separation
Previously, the location of seperation would try to find the area closest to a specified separation area.
However, this would include all the areas even before the location of minimum area. This means the
separation point could occur before the minimum area which does not make sense.

This has been fixed because the separation area now accounts for a weighting that ensures the separation
location is behind the minimum area point.