Refactor adjoint and fix coontrols implementation bugs
Refactored test_adjoint saving time in a few places where you only have
to compute the gradient once. Also tried to make it cleaner.

Previous code added a controls parameters for specifying the forward
simulation but it was not implemented fully since it was not written to
the simulation history. This should be fixed now.