Refactor solid solution/model and solver parameters
Moved the solid time step setting thing to a property so you don't have
to deal with the fact that the time step is variable in each cell (a
hack so derivatives can be taken)

Refactor solving of the solid into the solid model class.

Moved some commong solver parameter defaults into their own file.