Refactor time step to property and add test of FSAI model integration
Added a basic FSAI model that can be integrated forward in time. The
coupling has not been worked out properley yet, so that still has to be
fixed.

I also changed `dt` to a property that can be set for each model. I
think this makes the model class a bit cleaner.