Remove references to initial control and old time step setting methods
Only a single control for the given time step is needed so references to
a initial control have been removed. This makes the code a lot cleaner.

Also the set_time_step method has been refactored to a dt property on
each model. This is also a nicer way to set/access the time step in my
opinion.