Refactor functional to return valid state/property variables
Refactored functionals into solid, fluid, and coupled functionals.
These have an interface to get the dstate sensitivity whether a FSI or
FSAI model is passed in.

This also led to a bunch of refactor in the adjoint integration to make
it work generally with either FSI or FSAI models.

I still have to test if the FSAI model is working propertly though.