Refactor gradient code to general adjoint operation
`adjoint.adjoint` was previously a code that combined computations of
the adjoint of the model timestepping procedure with partial derivatives
of a functional to compte the gradient of the functional. Since the
adjoint operation itself is useful to use, this was refactored into an
adjoint operation, and an operation that computes gradients of a
functional by adding to the adjoint operation.