Change forward.integrate to accept a StateFile instance
forward.integrate now requires a StateFile instance rather than opening
a StateFile from a provided file path inside the function. This is
a nicer interface because the StateFile is often used again (for example
states are read to perform a linearized forward/adjoint integration) after
it is written to from forward.integrate.