Simplify how properties are specified for solid models
Previously solid properties were defined in two places: as a class
variable as well as in how they were exported from the form definitions.
This was simplified to be based only on the exported form definitions.

The way sensitivities of the residual with respect to model properties was
computed was also simplified.