Fix mistake in StrainRateDampingWork
The StrainRateDampingWork functional is not correct and will not work
the way it was written. It was repurposed to calculate strain work for
the approximate 3D case, but in the future all the damping works should
be generalized based on stress fuctions that are specified by the actual
solid model.

That would be a good way to generalize all the current functions to a
single class.