Fix bug in TransferWorkByVelocity
Transfer work from fluid done on solid was computed incorrectly because
of explicit vs implicity coupling. In the case of explicit coupling, the
pressure loading on the surface is based on the previous solid state so
care must be taken to update the previous and current states so that the
surface pressure is computed correctly.