Fix Bernoulli instabilities by bounding area function
Instabilities could occur due to differences between the `smoothmin`
computed minimum area, and the actual physical minimum areas based on
nodal positions. When absolute values of the area are small, the small
difference between the two areas can result in large area ratios which
leads to unphysicaly pressure variations according to Bernoulli's
formula. To fix this, the areas used to compute pressure in Bernoulli
where further bounded a minimum based on the `smoothmin` area.