Fix fluid smoothed separation model
The smoothed separation model was modified to include the limiting
discontinuous behaviour as a special case.

In addition, the previous separation model was found to give the wrong
behaviour in the limit of small smoothing parameters; instead of
selecting the point past the minimum area location closest to the
separation area, it would simply select the point closes to the
separation area which is sometimes located before the minimum area
location. This is not correct so the separation model was modified.