maint: Move assembly into `solve()` for performance
Previously `linear_subproblem` would assemble the system state when
called which is wasteful. This is because if the residual is evaluated
to satisfy convergence, then an iterative solver does not need to
perform a solve and therefore doesn't need to assemble the jacobian. The
previous method assemble the Jacobian everytime anyway.